requirements.txt
aspell-python-py3==1.15
spacy==2.2.4
fairseq==0.9.0
